+++ doc/changes.rst
@@ -5,13 +5,15 @@ desitarget Change Log
 2.6.2 (unreleased)
 ------------------
 
-* Option to read subset of columns from HEALPix-split MTLs [`PR #804`_].
+* Option to read subset of columns from HEALPix-split MTLs [`PR #805`_].
     * Specifically in :func:`io.read_mtl_in_hp()`.
-    * Partial fix for `issue #802`_.
+    * Addresses `issue #802`_.
+    * Partial fix for `issue #804`_.
     * Also address some deprecation warnings.
 
 .. _`issue #802`: https://github.com/desihub/desitarget/issues/802
-.. _`PR #804`: https://github.com/desihub/desitarget/pull/804
+.. _`issue #802`: https://github.com/desihub/desitarget/issues/804
+.. _`PR #804`: https://github.com/desihub/desitarget/pull/805
 
 2.6.1 (2023-08-23)
 ------------------